cli/requirements.txt
click>=8.1.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.34.0
google-generativeai>=0.8.0
rich>=13.7.0
pydantic>=2.6.0
requests>=2.31.0
gitpython>=3.1.0
pathspec>=0.12.0
keyboard>=0.13.5
questionary>=2.0.0
prompt-toolkit>=3.0.0
pygments>=2.17.0
watchdog>=3.0.0

# Additional LLM provider SDKs
cohere>=5.0.0
replicate>=0.25.0
huggingface_hub>=0.20.0
ai21>=2.0.0
boto3>=1.34.0

pyfiglet
